Fixes typo and avoids uncessary rotations
An `if` statement was added to avoid calling rotate_field if this is not needed (as the present implementation of rotate_field is not working, this is also the typical case).
The wrong old "BMF" acronym was substituted by "BFM". The previous function name still work, but a deprecation warning is issued when they are used.